2023 Korea Physics Undergraduate Course Network Subject co-occurrence Network
This is an undergraduate student project in the Network Science Laboratory, Department of Physics, Gyeongsang National University(GNU). We quantified the subject diversity and core-centricity of physics departments in Korea through network science. The red nodes represent the physics departments in operation in 2023, and the blue nodes represent the curriculum organized by each school. Of course, it is a bipartite network :)